+++ src/hooks/useTypeaheadQuery.ts
@@ -11,24 +11,21 @@ import { useDebouncedValue } from './useDebouncedValue';
 
 export interface UseTypeaheadQueryOptions {
   debounceTime?: number;
-  isQueryLoading?: boolean;
   onDebouncedChange?: (value: string) => void;
 }
 
 export interface UseTypeaheadQueryResult {
-  isLoading: boolean;
   query: string;
   setQuery: Dispatch<SetStateAction<string>>;
 }
 
 export const useTypeaheadQuery = ({
   debounceTime,
-  isQueryLoading,
   onDebouncedChange,
 }: UseTypeaheadQueryOptions): UseTypeaheadQueryResult => {
   const onDebouncedChangeFn = useRef(onDebouncedChange);
   const [query, setQuery] = useState('');
-  const { debouncedValue, isDebouncing } = useDebouncedValue<string>(
+  const { debouncedValue } = useDebouncedValue<string>(
     query,
     debounceTime ?? 400,
   );
@@ -40,7 +37,6 @@ export const useTypeaheadQuery = ({
     onDebouncedChangeFn.current?.(currentQuery);
   }, [currentQuery]);
   return {
-    isLoading: isQueryLoading || isDebouncing,
     query,
     setQuery,
   };